Code style updates
