+++ rasa/shared/core/events.py
@@ -1936,79 +1936,3 @@ class SessionStarted(AlwaysEqualEventMixin):
         """Applies event to current conversation state."""
         # noinspection PyProtectedMember
         tracker._reset()
-
-
-class WronglyPredictedAction(ActionExecuted):
-    """The model predicted the wrong action.
-
-    Mostly used to mark wrong predictions and be able to
-    dump them as stories."""
-
-    type_name = "wrong_action"
-
-    def __init__(
-        self,
-        action_name_target: Text,
-        action_text_target: Text,
-        action_name_prediction: Text,
-        policy: Optional[Text] = None,
-        confidence: Optional[float] = None,
-        timestamp: Optional[float] = None,
-        metadata: Optional[Dict] = None,
-        predicted_action_unlikely_intent: bool = False,
-    ) -> None:
-        """Creates event for a successful event execution.
-
-        See the docstring of the parent class `ActionExecuted` for more information.
-        """
-        self.action_name_prediction = action_name_prediction
-        self.predicted_action_unlikely_intent = predicted_action_unlikely_intent
-        super().__init__(
-            action_name_target,
-            policy,
-            confidence,
-            timestamp,
-            metadata,
-            action_text=action_text_target,
-        )
-
-    def inline_comment(self) -> Text:
-        """A comment attached to this event. Used during dumping."""
-        comment = f"predicted: {self.action_name_prediction}"
-        if self.predicted_action_unlikely_intent:
-            return f"{comment} after {ACTION_UNLIKELY_INTENT_NAME}"
-        return comment
-
-    def as_story_string(self) -> Text:
-        """Returns the story equivalent representation."""
-        return f"{self.action_name}   <!-- {self.inline_comment()} -->"
-
-    def __repr__(self) -> Text:
-        """Returns event as string for debugging."""
-        return (
-            f"WronglyPredictedAction(action_target: {self.action_name}, "
-            f"action_prediction: {self.action_name_prediction}, "
-            f"policy: {self.policy}, confidence: {self.confidence}, "
-            f"metadata: {self.metadata})"
-        )
-
-
-class WarningPredictedAction(ActionExecuted):
-    """The model predicted the correct action with warning."""
-
-    type_name = "warning_predicted"
-
-    def __init__(
-        self,
-        action_name_prediction: Text,
-        action_name: Optional[Text] = None,
-        policy: Optional[Text] = None,
-        confidence: Optional[float] = None,
-        timestamp: Optional[float] = None,
-    ):
-        self.action_name_prediction = action_name_prediction
-        super().__init__(action_name, policy, confidence, timestamp)
-
-    def inline_comment(self) -> Text:
-        """A comment attached to this event. Used during dumping."""
-        return f"predicted: {self.action_name_prediction}"
