+++ rasa/shared/core/training_data/structures.py
@@ -31,7 +31,6 @@ from rasa.shared.core.events import (
     ActionExecuted,
     Event,
     SessionStarted,
-    WronglyPredictedAction,
 )
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.exceptions import RasaCoreException
@@ -236,11 +235,6 @@ class StoryStep:
         return (
             type(event) == ActionExecuted
             and event.action_name == ACTION_UNLIKELY_INTENT_NAME
-        ) or (
-            type(event) == WronglyPredictedAction
-            and event.action_name
-            == event.action_name_prediction
-            == ACTION_UNLIKELY_INTENT_NAME
         )
 
     @staticmethod
