+++ tests/test_model_testing.py
@@ -12,12 +12,15 @@ import rasa.shared.utils.io
 import rasa.utils.io
 from rasa.core.agent import Agent
 from rasa.shared.core.events import UserUttered
-from rasa.core.test import EvaluationStore, WronglyClassifiedUserUtterance
+from rasa.core.test import (
+    EvaluationStore,
+    WronglyClassifiedUserUtterance,
+    WronglyPredictedAction,
+)
 from rasa.shared.core.trackers import DialogueStateTracker
 from rasa.shared.core.training_data.story_writer.yaml_story_writer import (
     YAMLStoryWriter,
 )
-from rasa.shared.core.events import WronglyPredictedAction
 import rasa.model
 import rasa.cli.utils
 from rasa.nlu.test import NO_ENTITY
